+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 As of 2022-06-28, the published versions is `0.4.22` for Chrome and `0.4.22` for
 Firefox.
 
+## Unreleased
+
+- General: don't fail if search result container isn't detected, just write to
+  console.
+
 ## 0.4.22
 
 - Google: fix navigation in some feature snippets