+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Firefox.
 ## Unreleased
 
 - Google: fix detecting more results when in the new "continuous results" experience
+- Google: fix search box keybindings (#515)
 - General: don't fail if search result container isn't detected, just write to
   console.
 - YouTube: fix navigation in non-grid pages