+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Firefox.
   console.
 - YouTube: fix navigation in non-grid pages
 - YouTube: fix grid navigation in home page (issue #405)
+- YouTube: debounce results loading calls
 
 ## 0.4.22