+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Firefox.
 - Google: fix detecting more results when in the new "continuous results" experience
 - Google: fix search box keybindings (#515)
 - Google: fix job cards (#406)
+- Google: exclude sidebar
 - General: don't fail if search result container isn't detected, just write to
   console.
 - YouTube: fix navigation in non-grid pages